refactor(whispering): move recording deletes into operations
